Reader: 32-bit pixel writer index, pictures fill the page, combined base for image pages
DirectPixelWriter computed the framebuffer byte index as uint16_t. Every
panel CrossPoint has run on had a framebuffer under 64 KB (X4: 60 x 800 =
48 KB); the e-Minimal 7.8" is 234 x 1404 = 328 KB, so every image pixel
below row 280 wrapped to the top of the frame and a cover came out as
pieces along the right edge. uint32_t.

Images never scaled up, only down to fit, so a cover sized for a 480 x 800
page was a 5 cm stamp on a 1404 x 1872 one. Pictures (>= 300 px each way)
without CSS sizing may now scale up to fit the container and viewport;
smaller images (icons, drop caps, ornaments) keep their size. The pixel
cache is built at display size, so an upscaled image is decoded and
dithered at that size, not stretched. Section cache v46 so cached layouts
rebuild.

Image pages join the combined-base path on the IT8951: it takes a whole
4bpp frame per load, so an image page combines exactly like a text page,
and the separate B/W base only cost a second 1.2 s refresh. A dithered
picture driven differentially leaves the heaviest residue the reader
shows, so the page after one is a clean refresh, as the separate-base path
already did.

One strip per plane on the IT8951 (329 KB scratch, PSRAM): each strip
re-walks the page with clipping, so at 80 rows a full-page picture was
walked seven times (~0.6 s each) and each AA plane cost ~2x the B/W
render. Now three passes; a full-page cover shows in ~4 s instead of 7-9.

Submodule freeink-sdk -> b82837b: the IT8951 driver's partial-area loads,
gray-aware clears, wake scrub and load-clock options.

+++ lib/Epub/Epub/converters/DirectPixelWriter.h
@@ -177,7 +177,9 @@ struct DirectPixelWriter {
     const int sy = phyY - originY;
     if (static_cast<unsigned>(sy) >= static_cast<unsigned>(clipRows)) return;
 
-    const uint16_t byteIndex = static_cast<uint16_t>(sy * displayWidthBytes + (phyX >> 3));
+    // 32-bit: a 1872x1404 frame is 328 KB, and a 16-bit index wrapped every
+    // image pixel below row 280 to the top of the frame (e-Minimal, 2026-09-18).
+    const uint32_t byteIndex = static_cast<uint32_t>(sy) * displayWidthBytes + static_cast<uint32_t>(phyX >> 3);
     const uint8_t bitMask = 1 << (7 - (phyX & 7));
 
     if (state) {
